refactor patient's QNR/qb_timeline state before & after functionality for better reuse.
