+++ manage.py
@@ -44,7 +44,11 @@ from portal.models.questionnaire_bank import (
     QuestionnaireBank,
     add_static_questionnaire_bank,
 )
-from portal.models.questionnaire_response import QuestionnaireResponse
+from portal.models.questionnaire_response import (
+    QuestionnaireResponse,
+    capture_patient_state,
+    present_before_after_state,
+)
 from portal.models.relationship import add_static_relationships
 from portal.models.research_study import (
     BASE_RS_ID,
@@ -802,71 +806,6 @@ def merge_users(src_id, tgt_id, actor):
         user_id=acting_user.id,
         subject_id=tgt_id)
     print(message)
-
-
-def capture_patient_state(patient_id):
-    """Call to capture QBT and QNR state for patient, used for before/after"""
-    qnrs = QuestionnaireResponse.qnr_state(patient_id)
-    tl = QBT.timeline_state(patient_id)
-    return {'qnrs': qnrs, 'timeline': tl}
-
-
-def present_before_after_state(user_id, external_study_id, before_state):
-    from portal.dict_tools import dict_compare
-    after_qnrs = QuestionnaireResponse.qnr_state(user_id)
-    after_timeline = QBT.timeline_state(user_id)
-    qnrs_lost_reference = []
-
-    def visit_from_timeline(qb_name, qb_iteration, timeline_results):
-        """timeline results have computed visit name - quick lookup"""
-        for visit, name, iteration in timeline_results.values():
-            if qb_name == name and qb_iteration == iteration:
-                return visit
-        raise ValueError(f"no visit found for {qb_name}, {qb_iteration}")
-
-    # Compare results
-    added_q, removed_q, modified_q, same = dict_compare(
-        after_qnrs, before_state['qnrs'])
-    assert not added_q
-    assert not removed_q
-
-    added_t, removed_t, modified_t, same = dict_compare(
-        after_timeline, before_state['timeline'])
-
-    if any((added_t, removed_t, modified_t, modified_q)):
-        print(f"\nPatient {user_id} ({external_study_id}):")
-    if modified_q:
-        print("\tModified QNRs (old, new)")
-        for mod in sorted(modified_q):
-            print(f"\t\t{mod} {modified_q[mod][1]} ==>"
-                  f" {modified_q[mod][0]}")
-            # If the QNR previously had a reference QB and since
-            # lost it, capture for reporting.
-            if (
-                    modified_q[mod][1][0] != "none of the above" and
-                    modified_q[mod][0][0] == "none of the above"):
-                visit_name = visit_from_timeline(
-                    modified_q[mod][1][0],
-                    modified_q[mod][1][1],
-                    before_state["timeline"])
-                qnrs_lost_reference.append((visit_name, modified_q[mod][1][2]))
-    if added_t:
-        print("\tAdditional timeline rows:")
-        for item in sorted(added_t):
-            print(f"\t\t{item} {after_timeline[item]}")
-    if removed_t:
-        print("\tRemoved timeline rows:")
-        for item in sorted(removed_t):
-            print(
-                f"\t\t{item} "
-                f"{before_state['timeline'][item]}")
-    if modified_t:
-        print(f"\tModified timeline rows: (old, new)")
-        for item in sorted(modified_t):
-            print(f"\t\t{item}")
-            print(f"\t\t\t{modified_t[item][1]} ==> {modified_t[item][0]}")
-
-    return after_qnrs, after_timeline, qnrs_lost_reference
 
 
 @app.cli.command()
